# api/openapi.yaml
openapi: 3.1.0
info:
  title: Conflow Local API
  version: 0.4.0
  description: |
    Versioned same-origin API between the embedded React GUI and the local Go service.
servers:
  - url: /api/v1
    description: Same-origin local Conflow service
paths:
...
  /project:
    get:
      operationId: getProject
      summary: Get the loaded project
      responses:
        "200":
          description: Project
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            ETag:
              $ref: "#/components/headers/ETag"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
    put:
      operationId: updateProject
      summary: Replace editable project metadata
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectInput"
      responses:
        "200":
          description: Updated project
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            ETag:
              $ref: "#/components/headers/ETag"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "412":
          $ref: "#/components/responses/ManifestRevisionMismatch"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
  /environments:
    get:
      operationId: listEnvironments
      summary: List project environments
      responses:
        "200":
          description: Environments
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            ETag:
              $ref: "#/components/headers/ETag"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentsResponse"
    post:
      operationId: createEnvironment
      summary: Create an environment
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnvironmentInput"
      responses:
        "201":
          description: Created environment
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            ETag:
              $ref: "#/components/headers/ETag"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/StateConflict"
        "412":
          $ref: "#/components/responses/ManifestRevisionMismatch"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
  /environments/{environment_id}:
    parameters:
      - $ref: "#/components/parameters/EnvironmentID"
    get:
      operationId: getEnvironment
      summary: Get an environment
      responses:
        "200":
          description: Environment
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            ETag:
              $ref: "#/components/headers/ETag"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateEnvironment
      summary: Replace an environment without changing its ID
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEnvironmentInput"
      responses:
        "200":
          description: Updated environment
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            ETag:
              $ref: "#/components/headers/ETag"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/ManifestRevisionMismatch"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
    delete:
      operationId: deleteEnvironment
      summary: Delete an environment
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "200":
          description: Deleted environment ID
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            ETag:
              $ref: "#/components/headers/ETag"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteEnvironmentResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/StateConflict"
        "403":
          $ref: "#/components/responses/Forbidden"
        "412":
          $ref: "#/components/responses/ManifestRevisionMismatch"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
...
  /packs/{pack_name}/versions/{pack_version}:
    parameters:
      - $ref: "#/components/parameters/PackName"
      - $ref: "#/components/parameters/PackVersion"
    get:
      operationId: getPack
      summary: Get a versioned configuration pack metadata
      responses:
        "200":
          description: Pack metadata
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            ETag:
              $ref: "#/components/headers/PackETag"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackMetadataResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /packs/{pack_name}/versions/{pack_version}/schema:
    parameters:
      - $ref: "#/components/parameters/PackName"
      - $ref: "#/components/parameters/PackVersion"
      - $ref: "#/components/parameters/SchemaVersion"
    get:
      operationId: getPackSchema
      summary: Get a configuration pack form schema
      responses:
        "200":
          description: Versioned, declarative pack schema
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            ETag:
              $ref: "#/components/headers/PackETag"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackSchemaResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/SchemaIncompatible"
components:
  parameters:
    EnvironmentID:
      name: environment_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z][a-z0-9-]{1,62}$"
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Quoted local manifest revision returned by ETag.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
    PackName:
      name: pack_name
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z][a-z0-9-]{0,62}$"
    PackVersion:
      name: pack_version
      in: path
      required: true
      schema:
        type: string
        pattern: "^v[1-9][0-9]*$"
    SchemaVersion:
      name: schema_version
      in: query
      required: false
      description: Exact schema version understood by the client. Omit to retrieve the current schema.
      schema:
        type: integer
        minimum: 1
  headers:
    CacheControl:
      schema:
        type: string
        const: no-store
    ETag:
      description: Quoted local manifest revision.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
    PackETag:
      description: Quoted in-process Pack registry revision; this is independent of the project manifest revision.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
    RequestID:
      schema:
        type: string
  responses:
    BadRequest:
      description: Malformed JSON, unknown fields, or invalid request headers.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Browser origin is not allowed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    StateConflict:
      description: Resource already exists or the current state forbids the operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ManifestRevisionMismatch:
      description: The manifest revision changed; the response contains the authoritative current manifest state from the same snapshot as the returned ETag.
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        ETag:
          $ref: "#/components/headers/ETag"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ManifestRevisionMismatchResponse"
          example:
            error:
              code: revision_mismatch
              message: The project manifest changed. Review the current state before saving again.
              request_id: req_01JEXAMPLE
              current_revision: 2
              current_state:
                project:
                  id: photo-editor
                  name: Photo Editor from Git
                  pack_ref: mobile-ad-monetization/v1
                  source_type: managed-file
                environments:
                  - id: development
                    name: Development
                    kind: development
                    provider:
                      type: firebase-remote-config
                      project_id: photo-editor-dev
                    publish:
                      requires_confirmation: false
    ValidationFailed:
      description: Request is structurally valid but violates project rules.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    SchemaIncompatible:
      description: The requested pack schema version is not compatible with this pack version.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnsupportedMediaType:
      description: Mutation body is not application/json.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    PreconditionRequired:
      description: If-Match is required for this mutation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ManifestState:
      type: object
      additionalProperties: false
      required: [project, environments]
      properties:
        project:
          $ref: "#/components/schemas/Project"
        environments:
          type: array
          items:
            $ref: "#/components/schemas/Environment"
    ProjectResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/Project"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    EnvironmentsResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Environment"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    EnvironmentResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/Environment"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    DeleteEnvironmentResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [deleted_id]
          properties:
            deleted_id:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    PackMetadataResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/PackMetadata"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    PackSchemaResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/PackSchema"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ResponseMeta:
      type: object
      additionalProperties: false
      required: [request_id, revision]
      properties:
        request_id:
          type: string
        revision:
          type: integer
          minimum: 1
    Project:
      type: object
      additionalProperties: false
      required: [id, name, pack_ref, source_type]
      properties:
        id:
          type: string
          pattern: "^[a-z][a-z0-9-]{1,62}$"
        name:
          type: string
          minLength: 1
          maxLength: 120
        pack_ref:
          type: string
        source_type:
          type: string
          enum: [managed-file, git-json]
    UpdateProjectInput:
      type: object
      additionalProperties: false
      required: [id, name]
      properties:
        id:
          type: string
          pattern: "^[a-z][a-z0-9-]{1,62}$"
        name:
          type: string
          minLength: 1
          maxLength: 120
    Environment:
      type: object
      additionalProperties: false
      required: [id, name, kind, provider, publish]
      properties:
        id:
          type: string
          pattern: "^[a-z][a-z0-9-]{1,62}$"
          description: Stable opaque environment identifier. It cannot be changed after creation and carries no environment semantics.
        name:
          type: string
          minLength: 1
          maxLength: 120
          description: Human-readable environment name shown in the UI.
        kind:
          $ref: "#/components/schemas/EnvironmentKind"
        provider:
          $ref: "#/components/schemas/ProviderConfig"
        publish:
          $ref: "#/components/schemas/PublishConfig"
      example:
        id: production
        name: Production
        kind: production
        provider:
          type: firebase-remote-config
          project_id: photo-editor-prod
        publish:
          requires_confirmation: true
    CreateEnvironmentInput:
      type: object
      additionalProperties: false
      required: [id, name, kind, provider, publish]
      properties:
        id:
          type: string
          pattern: "^[a-z][a-z0-9-]{1,62}$"
        name:
          type: string
          minLength: 1
          maxLength: 120
        kind:
          $ref: "#/components/schemas/EnvironmentKind"
        provider:
          $ref: "#/components/schemas/ProviderConfig"
        publish:
          $ref: "#/components/schemas/PublishConfig"
    UpdateEnvironmentInput:
      type: object
      additionalProperties: false
      required: [name, provider, publish]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        provider:
          $ref: "#/components/schemas/ProviderConfig"
        publish:
          $ref: "#/components/schemas/PublishConfig"
    EnvironmentKind:
      type: string
      enum: [development, staging, production, custom]
      description: Server-authoritative environment category. Only production activates the persistent Production UI state. It cannot be changed after creation.
    ProviderConfig:
      type: object
      additionalProperties: false
      required: [type, project_id]
      properties:
        type:
          type: string
          const: firebase-remote-config
        project_id:
          type: string
          minLength: 1
          maxLength: 128
    PublishConfig:
      type: object
      additionalProperties: false
      required: [requires_confirmation]
      properties:
        requires_confirmation:
          type: boolean
...
    PackMetadata:
      type: object
      additionalProperties: false
      required: [ref, name, version, description, capabilities, schema_version, entity_types]
      properties:
        ref:
          type: string
          pattern: "^[a-z][a-z0-9-]{0,62}/v[1-9][0-9]*$"
        name:
          type: string
        version:
          type: string
        description:
          type: string
        capabilities:
          type: array
          items:
            type: string
        schema_version:
          type: integer
          minimum: 1
        entity_types:
          type: array
          items:
            $ref: "#/components/schemas/EntityMetadata"
    EntityMetadata:
      type: object
      additionalProperties: false
      required: [name, label, description, id_rule, deletion_policy, environment_override_fields]
      properties:
        name:
          type: string
        label:
          type: string
        description:
          type: string
        id_rule:
          $ref: "#/components/schemas/IDRule"
        deletion_policy:
          type: string
          enum: [restrict, cascade, allow]
        environment_override_fields:
          type: array
          items:
            type: string
    IDRule:
      type: object
      additionalProperties: false
      required: [pattern, min_length, max_length]
      properties:
        pattern:
          type: string
        min_length:
          type: integer
          minimum: 1
        max_length:
          type: integer
          minimum: 1
    PackSchema:
      type: object
      additionalProperties: false
      required: [version, entities, migrations]
      properties:
        version:
          type: integer
          minimum: 1
        entities:
          type: array
          items:
            $ref: "#/components/schemas/EntitySchema"
        migrations:
          type: array
          items:
            $ref: "#/components/schemas/SchemaMigration"
    EntitySchema:
      type: object
      additionalProperties: false
      required: [name, fields]
      properties:
        name:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/FieldSchema"
    FieldSchema:
      type: object
      additionalProperties: false
      required: [name, type, required, default, sensitivity, ui, validation]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [string, boolean, integer, number, object, array, reference]
        required:
          type: boolean
        default: {}
        sensitivity:
          type: string
          enum: [public, sensitive]
        ui:
          $ref: "#/components/schemas/FieldUI"
        validation:
          $ref: "#/components/schemas/FieldValidation"
    FieldUI:
      type: object
      additionalProperties: false
      required: [label, description, control, group, order]
      properties:
        label:
          type: string
        description:
          type: string
        control:
          type: string
        group:
          type: string
        order:
          type: integer
          minimum: 0
        placeholder:
          type: string
    FieldValidation:
      type: object
      additionalProperties: false
      required: [enum]
      properties:
        enum:
          type: array
          items: {}
        min_length:
          type: integer
          minimum: 0
        max_length:
          type: integer
          minimum: 0
        minimum:
          type: number
        maximum:
          type: number
    SchemaMigration:
      type: object
      additionalProperties: false
      required: [from_version, to_version, description]
      properties:
        from_version:
          type: integer
          minimum: 1
        to_version:
          type: integer
          minimum: 1
        description:
          type: string
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
    ManifestRevisionMismatchResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ManifestRevisionMismatchError"
    ManifestRevisionMismatchError:
      type: object
      additionalProperties: false
      required: [code, message, request_id, current_revision, current_state]
      properties:
        code:
          type: string
          const: revision_mismatch
        message:
          type: string
        request_id:
          type: string
        current_revision:
          type: integer
          minimum: 1
          description: Revision of current_state and the ETag response header.
        current_state:
          $ref: "#/components/schemas/ManifestState"
    Error:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        current_revision:
          type: integer
          minimum: 1
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        path:
          type: string
        code:
          type: string
        message:
          type: string