# internal/server/api.go
package server

import (
	"errors"
	"net/http"
	"strconv"

	"github.com/ConteMan/conflow/internal/app"
	"github.com/ConteMan/conflow/internal/packs"
	"github.com/ConteMan/conflow/internal/project"
)

type api struct {
	service *app.Service
}

func newAPI(service *app.Service) *api {
	return &api{service: service}
}

func (a *api) handler() http.Handler {
	mux := http.NewServeMux()
	mux.HandleFunc("GET /api/v1/health", a.health)
	mux.HandleFunc("GET /api/v1/bootstrap", a.bootstrap)
	mux.HandleFunc("GET /api/v1/project", a.getProject)
	mux.HandleFunc("PUT /api/v1/project", a.updateProject)
	mux.HandleFunc("GET /api/v1/environments", a.listEnvironments)
	mux.HandleFunc("POST /api/v1/environments", a.createEnvironment)
	mux.HandleFunc("GET /api/v1/environments/{environment_id}", a.getEnvironment)
	mux.HandleFunc("PUT /api/v1/environments/{environment_id}", a.updateEnvironment)
	mux.HandleFunc("DELETE /api/v1/environments/{environment_id}", a.deleteEnvironment)
	mux.HandleFunc("GET /api/v1/packs", a.listPacks)
	mux.HandleFunc("GET /api/v1/packs/{pack_name}/versions/{pack_version}", a.getPack)
	mux.HandleFunc("GET /api/v1/packs/{pack_name}/versions/{pack_version}/schema", a.getPackSchema)
	mux.HandleFunc("/api/v1/project", methodNotAllowed)
	mux.HandleFunc("/api/v1/environments", methodNotAllowed)
	mux.HandleFunc("/api/v1/environments/{environment_id}", methodNotAllowed)
	mux.HandleFunc("/api/v1/packs", methodNotAllowed)
	mux.HandleFunc("/api/v1/packs/{pack_name}/versions/{pack_version}", methodNotAllowed)
	mux.HandleFunc("/api/v1/packs/{pack_name}/versions/{pack_version}/schema", methodNotAllowed)
	mux.HandleFunc("/api/", routeNotFound)
	return apiMiddleware(mux)
}

func methodNotAllowed(writer http.ResponseWriter, request *http.Request) {
	writeAPIError(writer, request, http.StatusMethodNotAllowed, "method_not_allowed", "请求方法不允许", 0)
}

func routeNotFound(writer http.ResponseWriter, request *http.Request) {
	writeAPIError(writer, request, http.StatusNotFound, "route_not_found", "API 路径不存在", 0)
}

func (a *api) health(writer http.ResponseWriter, request *http.Request) {
	snapshot, err := a.service.Snapshot(request.Context())
	if err != nil {
		writeAPIError(writer, request, http.StatusServiceUnavailable, "project_unavailable", "项目配置当前不可用", 0)
		return
	}
	writeJSON(writer, http.StatusOK, healthResponse{Status: "ok", ProjectID: snapshot.Manifest.Project.ID})
}

func (a *api) bootstrap(writer http.ResponseWriter, request *http.Request) {
	snapshot, err := a.service.Snapshot(request.Context())
	if err != nil {
		a.writeError(writer, request, err)
		return
	}
	data := bootstrapData{
		Project:      projectDTOFrom(snapshot.Manifest),
		Environments: environmentsDTOFrom(snapshot.Manifest.Environments),
		Capabilities: capabilitiesDTO{ProjectEdit: true, EnvironmentManage: true},
	}
	writeSuccess(writer, request, http.StatusOK, data, snapshot.Revision)
}

func (a *api) getProject(writer http.ResponseWriter, request *http.Request) {
	snapshot, err := a.service.Snapshot(request.Context())
	if err != nil {
		a.writeError(writer, request, err)
		return
	}
	writeSuccess(writer, request, http.StatusOK, projectDTOFrom(snapshot.Manifest), snapshot.Revision)
}

func (a *api) updateProject(writer http.ResponseWriter, request *http.Request) {
	revision, ok := requireRevision(writer, request)
	if !ok {
		return
	}
	var input updateProjectInput
	if err := decodeJSON(writer, request, &input); err != nil {
		writeRequestError(writer, request, err)
		return
	}
	snapshot, err := a.service.UpdateProject(request.Context(), revision, project.Project{ID: input.ID, Name: input.Name})
	if err != nil {
		a.writeError(writer, request, err)
		return
	}
	writeSuccess(writer, request, http.StatusOK, projectDTOFrom(snapshot.Manifest), snapshot.Revision)
}

func (a *api) listEnvironments(writer http.ResponseWriter, request *http.Request) {
	snapshot, err := a.service.Snapshot(request.Context())
	if err != nil {
		a.writeError(writer, request, err)
		return
	}
	writeSuccess(writer, request, http.StatusOK, environmentsDTOFrom(snapshot.Manifest.Environments), snapshot.Revision)
}

func (a *api) createEnvironment(writer http.ResponseWriter, request *http.Request) {
	revision, ok := requireRevision(writer, request)
	if !ok {
		return
	}
	var input createEnvironmentInput
	if err := decodeJSON(writer, request, &input); err != nil {
		writeRequestError(writer, request, err)
		return
	}
	environment, ok := input.toProject()
	if !ok {
		writeAPIError(writer, request, http.StatusUnprocessableEntity, "validation_failed", "项目配置不合法", 0)
		return
	}
	snapshot, environment, err := a.service.CreateEnvironment(request.Context(), revision, environment)
	if err != nil {
		a.writeError(writer, request, err)
		return
	}
	writeSuccess(writer, request, http.StatusCreated, environmentDTOFrom(environment), snapshot.Revision)
}

func (a *api) getEnvironment(writer http.ResponseWriter, request *http.Request) {
	snapshot, environment, err := a.service.GetEnvironment(request.Context(), request.PathValue("environment_id"))
	if err != nil {
		a.writeError(writer, request, err)
		return
	}
	writeSuccess(writer, request, http.StatusOK, environmentDTOFrom(environment), snapshot.Revision)
}

func (a *api) updateEnvironment(writer http.ResponseWriter, request *http.Request) {
	revision, ok := requireRevision(writer, request)
	if !ok {
		return
	}
	var input updateEnvironmentInput
	if err := decodeJSON(writer, request, &input); err != nil {
		writeRequestError(writer, request, err)
		return
	}
	publish, ok := input.Publish.toProject()
	if !ok {
		writeAPIError(writer, request, http.StatusUnprocessableEntity, "validation_failed", "项目配置不合法", 0)
		return
	}
	environmentID := request.PathValue("environment_id")
	replacement := project.Environment{
		ID:       environmentID,
		Name:     input.Name,
		Provider: input.Provider.toProject(),
		Publish:  publish,
	}
	snapshot, environment, err := a.service.UpdateEnvironment(request.Context(), revision, environmentID, replacement)
	if err != nil {
		a.writeError(writer, request, err)
		return
	}
	writeSuccess(writer, request, http.StatusOK, environmentDTOFrom(environment), snapshot.Revision)
}

func (a *api) deleteEnvironment(writer http.ResponseWriter, request *http.Request) {
	revision, ok := requireRevision(writer, request)
	if !ok {
		return
	}
	environmentID := request.PathValue("environment_id")
	snapshot, err := a.service.DeleteEnvironment(request.Context(), revision, environmentID)
	if err != nil {
		a.writeError(writer, request, err)
		return
	}
	writeSuccess(writer, request, http.StatusOK, deleteEnvironmentData{DeletedID: environmentID}, snapshot.Revision)
}

func (a *api) listPacks(writer http.ResponseWriter, request *http.Request) {
	snapshot := a.service.ListPacks(request.Context())
	writeSuccess(writer, request, http.StatusOK, packSummariesDTOFrom(snapshot.Definitions), snapshot.Revision)
}

func (a *api) getPack(writer http.ResponseWriter, request *http.Request) {
	definition, revision, err := a.service.GetPack(request.Context(), request.PathValue("pack_name"), request.PathValue("pack_version"))
	if err != nil {
		a.writeError(writer, request, err)
		return
	}
	writeSuccess(writer, request, http.StatusOK, packMetadataDTOFrom(definition), revision)
}

func (a *api) getPackSchema(writer http.ResponseWriter, request *http.Request) {
	requestedVersion, err := schemaVersionFrom(request)
	if err != nil {
		writeAPIError(writer, request, http.StatusBadRequest, "invalid_request", "schema_version 必须是正整数", 0)
		return
	}
	schema, revision, err := a.service.GetPackSchema(request.Context(), request.PathValue("pack_name"), request.PathValue("pack_version"), requestedVersion)
	if err != nil {
		a.writeError(writer, request, err)
		return
	}
	writeSuccess(writer, request, http.StatusOK, packSchemaDTOFrom(schema), revision)
}

func schemaVersionFrom(request *http.Request) (*uint64, error) {
	values, exists := request.URL.Query()["schema_version"]
	if !exists {
		return nil, nil
	}
	if len(values) != 1 || values[0] == "" {
		return nil, errors.New("invalid schema version")
	}
	raw := values[0]
	version, err := strconv.ParseUint(raw, 10, 64)
	if err != nil || version == 0 {
		return nil, errors.New("invalid schema version")
	}
	return &version, nil
}

func (a *api) writeError(writer http.ResponseWriter, request *http.Request, err error) {
	var mismatch *project.RevisionMismatchError
	switch {
	case errors.As(err, &mismatch):
		writeManifestRevisionMismatch(writer, request, mismatch.Snapshot)
	case errors.Is(err, project.ErrNotFound):
		writeAPIError(writer, request, http.StatusNotFound, "environment_not_found", "环境不存在", 0)
	case errors.Is(err, project.ErrAlreadyExists):
		writeAPIError(writer, request, http.StatusConflict, "environment_already_exists", "环境 ID 已存在", 0)
	case errors.Is(err, app.ErrLastEnvironment):
		writeAPIError(writer, request, http.StatusConflict, "last_environment", "项目必须至少保留一个环境", 0)
	case errors.Is(err, packs.ErrUnknownPack):
		writeAPIError(writer, request, http.StatusNotFound, "pack_not_found", "配置包不存在", 0)
	case errors.Is(err, packs.ErrUnknownVersion):
		writeAPIError(writer, request, http.StatusNotFound, "pack_version_not_found", "配置包版本不存在", 0)
	case errors.Is(err, packs.ErrSchemaIncompatible):
		writeAPIError(writer, request, http.StatusUnprocessableEntity, "schema_incompatible", "配置包 schema 版本不兼容", 0)
	case errors.Is(err, project.ErrInvalidManifest):
		writeAPIError(writer, request, http.StatusUnprocessableEntity, "validation_failed", "项目配置不合法", 0)
	default:
		writeAPIError(writer, request, http.StatusInternalServerError, "internal_error", "操作失败", 0)
	}
}